Terminos da Busca de Cliente, inicio da busca de produto

+++ src/visao/FrmClientes.java
@@ -46,7 +46,7 @@ public class FrmClientes extends JFrame {
 		Cliente cliente = new Cliente();
 		cliente.setCodigo(Integer.parseInt(table.getValueAt(linha, 0).toString()));
 		cliente.setNome(table.getValueAt(linha, 1).toString());
-		cliente.setNumCadNacional(Integer.valueOf(table.getValueAt(linha, 2).toString()));
+		cliente.setNumCadNacional(Long.valueOf(table.getValueAt(linha, 2).toString()));
 		cliente.setNumCadEstadual(Integer.valueOf(table.getValueAt(linha, 3).toString()));
 		cliente.setTelefone(Integer.valueOf(table.getValueAt(linha, 4).toString()));
 		cliente.setEndereco(table.getValueAt(linha, 5).toString());
